+++ exchange/src/lib.rs
@@ -104,7 +104,7 @@ impl<'a> From<&BybitOrder<'a>> for OrderMessages {
         // TODO: this `try_into` is very dangerous. It needs to be improved.
         let order = Order {
             order_id: src.order_id.to_string(),
-            order_link_id: usize::from_str(src.order_link_id).unwrap(),
+            order_link_id: u64::from_str(src.order_link_id).unwrap(),
             order_status: src.order_status.try_into().unwrap(),
             symbol: src.symbol.to_string(),
             side: src.side.try_into().unwrap(),
@@ -122,7 +122,7 @@ impl<'a> From<&Execution<'a>> for OrderMessages {
     fn from(src: &Execution) -> Self {
         let order = OrderExecution {
             order_id: src.order_id.to_string(),
-            order_link_id: usize::from_str(src.order_link_id).unwrap(),
+            order_link_id: u64::from_str(src.order_link_id).unwrap(),
             qty: f64::from_str(src.exec_qty).unwrap(),
             price: f64::from_str(src.exec_price).unwrap(),
             remaining_qty: f64::from_str(src.leaves_qty).unwrap(),
@@ -141,7 +141,7 @@ pub struct OrderBuilder {
 }
 impl OrderBuilder {
     // TODO: should it be converted to an Into trait of `OrderMessages`?
-    pub fn build(self, order_id: String, order_link_id: usize) -> OrderMessages {
+    pub fn build(self, order_id: String, order_link_id: u64) -> OrderMessages {
         let order = Order {
             order_id,
             order_link_id,
@@ -163,7 +163,7 @@ impl OrderBuilder {
 pub struct OrderAmendedBuilder {
     pub symbol: String,
     pub order_id: String,
-    pub order_link_id: usize,
+    pub order_link_id: u64,
     pub qty: f64,
     pub price: String,
     pub new_price: bool,
@@ -194,7 +194,7 @@ impl OrderAmendedBuilder {
 #[derive(Clone, Serialize, Deserialize, Debug, Default)]
 pub struct Order {
     pub order_id: String,
-    pub order_link_id: usize,
+    pub order_link_id: u64,
     pub order_status: OrderStatus,
     pub symbol: String,
     pub side: OrderSide,
@@ -210,7 +210,7 @@ pub struct Order {
 #[derive(Clone, Serialize, Deserialize, Debug)]
 pub struct OrderExecution {
     pub order_id: String,
-    pub order_link_id: usize,
+    pub order_link_id: u64,
     // NOTE: price is the execution price
     pub price: f64,
     // NOTEL qty is the size of the execution
